feat (AR): adjust tutorial only to find the plane and then find the object
